+++ src/org/gnome/split/core/splitter/DefaultSplitEngine.java
@@ -180,7 +180,7 @@ public abstract class DefaultSplitEngine extends DefaultEngine
     /**
      * Notify the view that a part of the file has been read.
      */
-    protected void fireEngineDone(double done, double total) {
+    protected void fireEngineDone(long done, long total) {
         app.getEngineListener().engineDone(done, total);
     }
 
@@ -222,7 +222,7 @@ public abstract class DefaultSplitEngine extends DefaultEngine
             total += buffer.length;
 
             // Notify the view
-            this.fireEngineDone((double) total, (double) file.length());
+            this.fireEngineDone(total, file.length());
         }
 
         // Success
